Added javafx-maven-plugin to pom
I’ve added javafx-maven-plugin to the project’s pom for experimental
purposes. The javafx-maven-plugin allows the building of fully native,
JRE-independent applications, which may be very attractive in the
future. At the moment, however, it can be frustrating to use (it can
only build natives for the platform you’re currently on), and produces
huge binaries (Topsoil swells from less than a meg to 60 or 70 MB). But
it shouldn’t interfere with the usual build process at all and seems
useful to at least have available in the codebase.