Removed ConvertFX from the repo
Removed ConvertFX from the repo in order to remove external dependencies
and simplify the codebase in preperation for the upcoming refactoring.

The internal NodeToSVGConverter was patched to allow it to correctly
handle clip paths.

In the upcoming months, the plan is to replace all of this with
JavaScript charts, and we'll then be able to produce SVGs with 100%
fidelity.