Added FindBugs plugin and removed old charts
Added the FindBugs plugin in order to enforce higher code quality
standards.

As part of adding FindBugs, classes that are not used by the JavaScript
charts were removed. Additionally, all existing FindBugs warnings were
corrected.

Slight improvements to the build script were also made. Most notably,
the gradle.properties file was removed, so that Gradle no longer
defaults to the daemon. This is because the daemon was causing the
installBuster task to fail.

A bug was found, corrected, and tested for in
GetApplicationDirectoryOperation.

Going forward, all contributions to the project should not produce any
FindBugs warnings. This policy will remain in place until it becomes
absolutely necessary to relax it.