Tweaked Travis CI buildscript
Tweaked the Travis CI buildscript in an attempt to make TestFX more
consistent between developer machines and the CI server.

A test was removed that involved a file chooser. File choosers throw
errors with the new configuration.

A test was added to demonstrate the the new configuration is an
improvement upon the old one.