Improved generic type on `NumberField`
Changed superclass of `NumberField` from `Field<Number>` to
`Field<Double>`. The motivation for this was to allow the following
statement to compile:

    new SimpleEntry().set(new NumberField("Test Field"), 3.07);

Prior to this commit, you would need to cast the `NumberField` to a
`Field<Double>` in the example above.

Tests were added to ensure this and other basic properties of
`NumberField` hold.