Moved datasets into SQLite
Moved datasets into SQLite in order to leverage technologies like
SQLite, MyBatis, and Flyway. In this update, old datasets ARE NOT
automatically migrated into the new system and must be added
manually instead.

Tests in EntryTableColumnTest were disabled after proving to be
difficult to debug. Visually, this functionality appears to still
work. This should be troubleshooted again at a later date.