Split into app and core
Split the Topsoil project into two submodules, `app` and `core`, which
should provide better separation of concerns and allow other projects to
depend on *just* the core of Topsoil.

It should be noted that this commit removes OSXIncompatible, as it was
unused and didn't seem to belong in either subproject.