Added ability to delete datasets
Fixed #170 by adding the ability to delete datasets. This feature may be
accessed by navigating to the target dataset in the "Datasets" menu and
using that dataset's submenu.

As part of this change, under the hood, MyBatis now returns a dataset's
SQL id in a hidden field. This can be used to facilitate other updates
as well.