Added Travis CI addon to provide newer Java
Added the Travis CI apt addon. This allows the inclusion of the package
`oracle-java8-installer`, which should provide the latest version of
Java 8.

ControlsFX has been removed as a dependency, because the version of Java
installed by the addon, 8u60, has the JavaFX Dialogs API as part of the
standard library. This fixes #69.