Revamped chart properties
Rewrote much of the chart properties/settings stuff, hoping to get to
something stable that handles all of the necessary cases without being
too fancy. Currently, it's rough, but it's a start.

The biggest changes are (1) the change in paradigm that properties
(formerly settings) should never be set by the chart and (2) charts can
provide a displayable panel to configure themselves. The first change is
motivated by performance and simplicity. For example, charts shouldn't
and don't need to change their own title, so why were they able to? This
means that the plot window is not specified by properties. The second
change is a nice trick I saw while studying IntelliJ.